UI_Proj_basic/requirements.txt
streamlit==1.42.2
google-generative-ai==0.7.0
python-dotenv==1.0.1
requests==2.32.3
numpy==2.2.3
pandas==2.2.3 